Move HttpRequest node after refactoring
